Create Subscription.c
Pavan Wants to conduct a lecture for which he needs to set up an online meeting of exactly X minutes.The meeting paltform supports a meeting of 40 minutes without subscription and a meeting of unlimited duration with subscription. Determine whether Pavan need to take a subscription or not for setting the meet.